requirements.txt
asgiref==3.8.1
Django==5.2
dnspython==2.7.0
pymongo==4.12.0
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
